Refactor to make it work under RakuAST
- Add dependency to Slangify to abstract the slangification in a
  central place, to allow for possible future changes behind the
  scenes.
- Add preliminary support to allow running under RakuAST.  This
  requires a post 2023.09 version of Rakudo, specifically 9b29fdcab61
- Add some basic tests, also in "do" context
- Add .gitignore for convenience
- Does NOT bump version yet

For the record: thank you for creating this slang after the discussions
after Damian's blog post, now 4 years ago.  If you think it would be
better if this were a Raku community module, transfer the repo to
raku-community-modules, and I will take care of it there.  Of course,
if you would like to keep taking care of this module, that's also ok!